requirements.txt
customtkinter==5.2.2
sympy==1.12
pillow==10.3.0
matplotlib==3.8.4